Update tutorials-en.md
This blog post literally saved me. I'm hosting wordpress on an AWS lightsail instance. The only way to update php in this setting is to create a brand new instance and migrate all wordpress data to the new instance. This blog describes it perfectly. I followed the instructions and had zero issues. I hope that this can help others as well.